% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_DiffPar.R
\name{CP_DiffPar}
\alias{CP_DiffPar}
\title{Title Differential Rhythm Parameter Test}
\usage{
CP_DiffPar(
  x,
  Par = c("A"),
  TOJR = NULL,
  alpha = 0.05,
  p.adjust.method = "BH",
  parallel.ncores = 1
)
}
\arguments{
\item{x}{outpuf of CP_Rhythmicity(x1, x2)}

\item{Par}{charater string. One of "A", "phase", "M", "A&phase", or "A&phase&M"}

\item{TOJR}{toTOJR output. If NULL, rhythm.joint object in x will be used.}

\item{alpha}{cutoff of p-values for significant differential parameter shift.}

\item{p.adjust.method}{input for p.adjust() in R package \code{stat}}

\item{parallel.ncores}{integer. Number of cores used if using parallel computing with \code{mclapply()}. Not functional for windows system.}
}
\value{
A dataframe of differential rhythm parameter test results.
}
\description{
This function performs differential rhythm parameter tests.
}
\examples{
x = CP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)
rhythm.res = CP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
rhythm.diffPar = CP_DiffPar(rhythm.res, Par = "A&phase")
}
