% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_DiffRhythmicity.R
\name{CP_DiffRhythmicity}
\alias{CP_DiffRhythmicity}
\title{Title}
\usage{
CP_DiffRhythmicity(
  x1 = data1.rhythm,
  x2 = data2.rhythm,
  x.joint = joint.rhythm,
  period = 24,
  diffPar.method = "circacompare",
  diffPar.adjust = adj_minP,
  diffR2.method = "LR",
  nPermutation = 1000,
  permutation.save = getwd(),
  permutation.file.label = "Group1",
  alpha = 0.05,
  p.adjust.method = "BH",
  parallel = FALSE,
  cores = 5
)
}
\arguments{
\item{x1}{From gorup 1: a list containing the following component: data: dataframe or matrix with rows of genes and columns of samples; tod: time of death or time of expression corresponding to the columns in data component; label: the gene names or other labels of the genes}

\item{x2}{From group 2.}

\item{x.joint}{A list with two components: Rhythmic.Both contains the labels of genes that are rhythmic in both regions; Rhythmic.GT.One contains the labels of genes that are rhythmic in at least one region}

\item{period}{The length of the rhythmicity cycle. When it is 24, the signal is circadian.}

\item{diffPar.method}{A string. One of "two_cosinor1", "two_cosinor2", "circacompare", "permutation". The method "two_cosinor1" will test the overlap of two CI, and method "two_cosinor2" test if the estimate of group 2 is in CI of group 1.}

\item{diffPar.adjust}{FDR controlling function, one of adj_minP, adj_BH_separate, adj_BH_pool, adj_AWFisher}

\item{diffR2.method}{A string. One of "LR", "permutation", "bootstrap"}

\item{nPermutation}{A numeric value. Number of permutation performer, only required when method = "permutation". The smallest possible p-value in the result will be 1/(n_gene*nPermutation), but larger nPermutation takes longer computing time.}

\item{permutation.save}{A local directory where you want to save the permutation result for future use. If "NULL" then no result will be saved.}

\item{permutation.file.label}{pecial label for file name of permutation if you want to save the permutation file}

\item{alpha}{number between 0 to 1. The critical level used in differential parameter test.}

\item{p.adjust.method}{choose valid input for p.adjust.method by checking p.adjust.methods for p.adjust() in R package \code{stat}}

\item{parallel}{TRUE/FALSE. If TRUE, parallel computing using mclapply will be used, which does not work on windows system.}

\item{cores}{number of cores used if using parallel computing}
}
\value{
A list with two components: diffPar: result of differential parameter test for genes rhythmic in both groups; diffR2: result of differential R2 test for genes rhythmic in at least one group
}
\description{
Title
}
\examples{
#x1.time = runif(20, min = 0, max = 24)
#m1 = rnorm(100, 5); A1 = rnorm(100, 3); phase1 = runif(100, min = 0, max = 2*pi); sigma = 1
#noise.mat1 = matrix(rnorm(100*20, 0, sigma), ncol = 20, nrow = 100)
#signal.mat1 = t(sapply(1:100, function(a){m1[a]+A1[a]*cos(2*pi/24*x1.time+phase1[a])}))

#x2.time = runif(20, min = 0, max = 24)
#m2 = m1; A2 = 1.5*A1; phase2 = runif(100, min = 0, max = 2*pi); sigma = 1
#noise.mat2 = matrix(rnorm(100*20, 0, sigma), ncol = 20, nrow = 100)
#signal.mat2 = t(sapply(1:100, function(a){m2[a]+A2[a]*cos(2*pi/24*x2.time+phase2[a])}))

#x1 = list(data = noise.mat1 + signal.mat1, tod = x1.time, label = paste("gene", seq_len(100)))
#x1.Rhythm = CP_Rhythmicity(x1, parallel = FALSE)
#x2 = list(data = noise.mat2 + signal.mat2, tod = x2.time, label = paste("gene", seq_len(100)))
#x2.Rhythm = CP_Rhythmicity(x2, parallel = FALSE)

#x.joint = CP_JointRhythmicity(x1.Rhythm, x2.Rhythm, method = "AW-Fisher",
#para.list = list(param = "pvalue", range = "less", value = 0.05))

#diffRhythm.res = CP_DiffRhythmicity(x1.Rhythm, x2.Rhythm, x.joint)

}
