% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_GraphFunctions.R
\name{CP_PlotDisplay}
\alias{CP_PlotDisplay}
\title{Title Displaying CP_ScatterPlot outputs}
\usage{
CP_PlotDisplay(
  x = CP_ScatterPlot(x, genes.plot = NULL, Info1 = "gI", Info2 = "gII", filename =
    NULL, height = 8, width = 8),
  id = 1:2
)
}
\arguments{
\item{x}{CP_ScatterPlot output}

\item{id}{integer. Indexes for plots to display.}
}
\description{
Title Displaying CP_ScatterPlot outputs
}
