% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_PlotFittedCosinor.R
\name{CP_PlotFittedCosinor}
\alias{CP_PlotFittedCosinor}
\title{Title}
\usage{
CP_PlotFittedCosinor(
  x1 = x1.rhythm,
  x2 = x2.rhythm,
  genes.plot = c("CLOCK"),
  Info1 = "data1",
  Info2 = "data2",
  category1 = NULL,
  category1.label = NULL,
  category2 = NULL,
  category2.label = NULL,
  filename = NULL,
  height = 8,
  width = 8
)
}
\arguments{
\item{x1}{output of CP_Rhythmicity(). A list of data, tod, label and rhythmicity parameter estimtggplot2::aes}

\item{x2}{output of CP_Rhythmicity(). If NULL, then only one plot will be made}

\item{genes.plot}{labels of genes that wish to be plotted. Labels must be consistent with those given in x1 or x1 and x2. If NULL, the top 10 most rhythmic genes from x1 will be used}

\item{Info1}{used in title of the plot}

\item{Info2}{used in title of the plot}

\item{category1}{category of x1 if there are subgroups of samples. Must be a vector whose length equals number of samples in x1.}

\item{category1.label}{will be used as legend title}

\item{category2}{category of x2 if there are subgroups of samples. Must be a vector whose length equals number of samples in x2.}

\item{category2.label}{will be used as legend title}

\item{filename}{filename for image output. If NULL, only ggplot object will be returned.}

\item{height}{height of the output plot}

\item{width}{width of the output plot}
}
\value{
A list which contains: a list of ggplot2 objects if x2 is null; two lists if x2 is not null.
}
\description{
Title
}
\examples{
#Simulate two time series data with 20 time points and 100 genes

x1.time = runif(20, min = 0, max = 24)
m1 = rnorm(100, 5); A1 = rnorm(100, 3); phase1 = runif(100, min = 0, max = 2*pi); sigma1 = 1
noise.mat1 = matrix(rnorm(100*20, 0, sigma1), ncol = 20, nrow = 100)
signal.mat1 = t(sapply(1:100, function(a){m1[a]+A1[a]*cos(2*pi/24*x1.time+phase1[a])}))
x1 = list(data = as.data.frame(noise.mat1 + signal.mat1),
tod = x1.time,
label = paste("gene", seq_len(100)))
x1.rhythm = CP_Rhythmicity(x1, parallel = FALSE)

x2.time = runif(20, min = 0, max = 24)
m2 = m1 + 2; A2 = A1+1; phase2 = phase1+pi/4; sigma = 1
noise.mat2 = matrix(rnorm(100*20, 0, sigma1), ncol = 20, nrow = 100)
signal.mat2 = t(sapply(1:100, function(a){m2[a]+A2[a]*cos(2*pi/24*x2.time+phase2[a])}))
x2 = list(data = as.data.frame(noise.mat2 + signal.mat2),
tod = x2.time,
label = paste("gene", seq_len(100)))
x2.rhythm = CP_Rhythmicity(x2, parallel = FALSE)

CP_PlotFittedCosinor(x1.rhythm, x2.rhythm, genes.plot = NULL, Info1 = "data1", Info2 = "data2")
}
