% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_GraphFunctions.R
\name{CP_PlotPhase}
\alias{CP_PlotPhase}
\title{Title Circos Plots for Phase Summary}
\usage{
CP_PlotPhase(
  x,
  type,
  sig.cut = list(param = "pvalue", fun = "<", val = "0.05"),
  color.df = NULL,
  Info1 = "groupI",
  Info2 = "groupII",
  filename = "NULL",
  file.width = 8,
  file.height = 8,
  phase.start = -6,
  concordance.ref = 4,
  single.binwidth = 1,
  cir.x.breaks = seq(-12, 12, 4),
  cir.y.breaks = seq(-6, 18, 4),
  axis.text.size = 12,
  legend.position = "right"
)
}
\arguments{
\item{x}{output from CP_DiffPar()}

\item{type}{character string. Choices are "square", "circos_connected", "circos_diff", "circos_hist", "circos_hist2".}

\item{sig.cut}{A list. For single group plot, only genes satisfying sig.cut will be plotted; for two-group plots, all RhyBoth genes will be plotted and genes satisfying sig.cut will be plotted in a different color. The list has three components: \itemize{
\item param parameter used for the cutoff. Should be a column in x.
\item fun character string. Either "<", or ">"
\item val numeric. Number used for the cutoff}}

\item{color.df}{data frame for color coding. Only used for "circos_diff'. Only one of color.df and sig.cut should be specified. The data frame should have two columns named color and label. label will be displayed as legend. The color vector should have the same order as genes in x$diffPar.tab}

\item{Info1}{character string. Used in the plot title for group I.}

\item{Info2}{character string. Used in the plot title for group II (if exist).}

\item{filename}{character string of the filename for exporting. If NULL, plots are not saved.}

\item{file.width}{height of the export plot}

\item{file.height}{width of the export plot}

\item{phase.start}{numeric. What time among the periodic clock do you want the phase start? Default is -6, which is midnight if time is in ZT scale.}

\item{concordance.ref}{numeric. The radius where the concordance reference be plotted away from \eqn{\Delta}peak = 0.}

\item{single.binwidth}{numeric. The binwidth for plotting peak histogram. Used in "circos_hist" and "square".}

\item{cir.x.breaks}{numeric. A vector for breaks for the radius (peak difference) used in "circos_diff". Should only contains values from -period/2 to period/2 and it is recommended that the break is equal spaced.}

\item{cir.y.breaks}{numeric. A vector for breaks for the angles. Should start with phase.start and end with phase.start+period}

\item{axis.text.size}{numeric. Size for the axis text.}

\item{legend.position}{"left”,“top”, “right”, “bottom”, or "none"}
}
\description{
Title Circos Plots for Phase Summary
}
\examples{
x = CP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/2, pi*3/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)

rhythm.res = CP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
#notice that for CP_PlotPhase() the parameters tested by CP_DiffPar must contain "phase"
rhythm.diffPar = CP_DiffPar(rhythm.res, "phase")
#make scatter plots of \eqn{\delta}peak vs. peak in group I
CP_PlotPhase(rhythm.diffPar, "circos_diff")
#make a connected phase plot
CP_PlotPhase(rhythm.diffPar, "circos_connected")
}
