% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_ResidualDiagnostics.R
\name{CP_ResidualDiagnostics}
\alias{CP_ResidualDiagnostics}
\title{Title}
\usage{
CP_ResidualDiagnostics(
  x = x1.rhythm,
  genes.diag = NULL,
  plot = TRUE,
  filename = NULL,
  height = 8,
  width = 8
)
}
\arguments{
\item{x}{the output fro CP_Rhythmicity}

\item{genes.diag}{A vector of genes that wish to do residual diagnostics. If NULL, top 10 most rhythmic genes will be selcted.}

\item{plot}{If TRUE, plots will be output. Otherwise, only ggplots objects are returned.}

\item{filename}{If NULL, no file will be saved. If input a valid filename, the residual plots will be saved with the filename.}

\item{height}{height for the saved file. Only used if filename is not NULL.}

\item{width}{width for the saved file. Only used if filename is not NULL.}
}
\value{
A list with one list for each gene.
}
\description{
Title
}
\examples{

x1.time = runif(20, min = 0, max = 24)
m1 = rnorm(100, 5); A1 = rnorm(100, 3); phase1 = runif(100, min = 0, max = 2*pi); sigma1 = 1
noise.mat1 = matrix(rnorm(100*20, 0, sigma1), ncol = 20, nrow = 100)
signal.mat1 = t(sapply(1:100, function(a){m1[a]+A1[a]*cos(2*pi/24*x1.time+phase1[a])}))
x1 = list(data = as.data.frame(noise.mat1 + signal.mat1),
          tod = x1.time,
          label = paste("gene", seq_len(100)))
x1.Rhythm = CP_Rhythmicity(x1, parallel = FALSE)
CP_ResidualDiagnostics(x1.Rhythm)

}
