% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_GraphFunctions.R
\name{CP_ScatterPlot}
\alias{CP_ScatterPlot}
\title{Title Scatter Plots of Expression vs. Time.}
\usage{
CP_ScatterPlot(
  x,
  genes.plot = NULL,
  Info1 = "gI",
  Info2 = "gII",
  filename = NULL,
  file.width = 8,
  file.height = 8
)
}
\arguments{
\item{x}{CP_Rhythmicity() output}

\item{genes.plot}{a vector of character strings. Names of genes to be plotted. Names should be same in nomenclature as gname in the data list. If NULL, the top 10 most rhythmic genes from group I will be used}

\item{Info1}{character string. Used in the plot title for group I.}

\item{Info2}{character string. Used in the plot title for group II (if exist).}

\item{filename}{character string of the filename for exporting. If NULL, plots are not saved.}

\item{file.width}{height of the export plot}

\item{file.height}{width of the export plot}
}
\value{
A list of ggplot2 plots.
}
\description{
Title Scatter Plots of Expression vs. Time.
}
\examples{
x = CP_sim_data(ngene=1000, nsample=30, A1=c(1, 3), A2=c(1, 3),
phase1=c(0, pi/4), phase2=c(pi/4, pi/2),
M1=c(4, 6), M2=c(4, 6), sigma1=1, sigma2=1)

rhythm.res = CP_Rhythmicity(x1 = x[[1]], x2 = x[[2]])
rhythm.plots = CP_ScatterPlot(rhythm.res)
#to display plot in Rstudio use CP_PlotDisplay()
#display the first five plots
CP_PlotDisplay(rhythm.plots, id = 1:5)
}
