% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCosCurve.R
\name{fitCosCurve}
\alias{fitCosCurve}
\title{Fit cosinor function using optimization}
\usage{
fitCosCurve(
  tod = tt,
  y = yy,
  period = 24,
  parStart = list(amp = 3, phase = 0, offset = 0)
)
}
\arguments{
\item{tod}{a numeric vector. Time of death (e.g. time of expression of the gene). length should be the same as y}

\item{y}{a numeric vector. Gene expression level.}

\item{period}{period The length of the rhythmicity cycle. When it is 24 (default), the signal is circadian.}

\item{parStart}{Initial value for optimization purpose.}
}
\value{
A list of A, phase, M, peak, B1, B2, SST, SSE, R2.
Formula 1: \eqn{y = M + A * cos(2\pi/period * t+phase)}
Formula 2: \eqn{y = M + B1 * cos(2\pi/period * t) + B2 * cos(2*\pi/period * t) }
{A}{Amplitude based on formula 1.}
{phase}{Phase based on formula 1, phase is restricted within (0, 2\eqn{\pi}).}
{M}{Basal level (vertical shift) based on formula 1 or on formula 2.}
{B1}{B1 based on formula 2.}
{B2}{B2 based on formula 2.}
{tss}{Total sum of square.}
{rss}{Residual sum of square, SSE/n is the MLE of the variance sigma square.}
{R2}{Pseudo R2 defined as (tss - rss)/tss.}
}
\description{
Fit cosinor function
}
\details{
This function fit the expression and time to this function \deqn{y = M+A(cos(2\pi/24*t+phase))+\epsilon}, where no distribution assumption is made on \eqn{\epsilon}.
}
\examples{
set.seed(32608)
n <- 50
tt <- runif(n,0,24)
Amp <- 2
Phase <- 6
Offset <- 3
yy <- Amp * cos(2*pi/24 * tt + Phase) + Offset + rnorm(n,0,1)
fitCosCurve(tt, yy)
}
