% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_cosinor_OLS.R
\name{one_cosinor_OLS}
\alias{one_cosinor_OLS}
\title{Title Fit one-component cosinor with OLS for one gene}
\usage{
one_cosinor_OLS(
  tod = time,
  y = y,
  alpha = 0.05,
  period = 24,
  CI = TRUE,
  CI.type = "conservative"
)
}
\arguments{
\item{tod}{a numeric vector. Time of death (e.g. time of expression of the gene). length should be the same as y}

\item{y}{a numeric vector. Gene expression level.}

\item{alpha}{a number between 0 to 1. The critical level for the confidence interval}

\item{period}{The length of the rhythmicity cycle. When it is 24 (default), the signal is circadian.}

\item{CI}{If return CI}

\item{CI.type}{"conservative": a convervative CI will be returned; "delta": CI calculated with sd derived with delta method and the plug in estimate will be returned.}
}
\value{
{M}{A list of OLS estimate of M and the lower and upper limit of M corresponding to the critical level alpha}

{A}{A list of estimate of A, sd of A and the lower and upper limit of A corresponding to the critical level alpha}

{phase}{A list of estimate of phase, sd of phase and the lower and upper limit of phase corresponding to the critical level alpha}

{peak}{The conversion from phase to peak is  \eqn{peak = (period-period*phase/(2*\pi))}.}

{test}{A list of test result: F statistics, p-value and R2, \eqn{\sigma^2} (the variance of the error term)}
}
\description{
This function fit the expression and time to this function \deqn{y = M+A(cos(2\pi/24*t+phase))+\epsilon}, where \eqn{\epsilon \sim N(0, \sigma^2)}.
}
\examples{
#Simulate a time series data with 20 time points

x.time = runif(20, min = 0, max = 24)
m = 5; A = 3; phase = pi/4; sigma = 1
y = m+A*cos(2*pi/24*x.time+phase)+rnorm(20, 0, sigma)
est = one_cosinor_OLS(tod = x.time, y = y, alpha = 0.05, period = 24)

}
