% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_cosinor_OLS.R
\name{two_cosinor_OLS}
\alias{two_cosinor_OLS}
\title{Title Fit two-component cosinor with OLS for one gene}
\usage{
two_cosinor_OLS(
  tod = time,
  y = y,
  group,
  alpha = 0.05,
  period = 24,
  CI.type = "delta"
)
}
\arguments{
\item{tod}{a numeric vector. Time of death (e.g. time of expression of the gene). length should be the same as y}

\item{y}{a numeric vector. Gene expression level.}

\item{group}{A vector indicating the group label. 0 for group 1 and 1 for group 2.}

\item{alpha}{a number between 0 to 1. The critical level for the confidence interval}

\item{period}{The length of the rhythmicity cycle. When it is 24 (default), the signal is circadian.}

\item{CI.type}{"conservative": a convervative CI will be returned; "delta": CI calculated with sd derived with delta method and the plug in estimate will be returned.}
}
\value{
{M1}{A list of OLS estimate of M1 and the lower and upper limit of M1 corresponding to the critical level alpha}

{A1}{A list of estimate of A1, sd of A1 and the lower and upper limit of A1 corresponding to the critical level alpha}

{phase1}{A list of estimate of phase1, sd of phase1 and the lower and upper limit of phase1 corresponding to the critical level alpha}

{M2}{A list of OLS estimate of M2 and the lower and upper limit of M2 corresponding to the critical level alpha}

{A2}{A list of estimate of A2, sd of A2 and the lower and upper limit of A2 corresponding to the critical level alpha}

{phase2}{A list of estimate of phase2, sd of phase2 and the lower and upper limit of phase2 corresponding to the critical level alpha}

{test}{A list of test result: F statistics, p-value and R2, \eqn{\sigma^2} (the variance of the error term)}
}
\description{
This function fit the expression and time to this function \deqn{y = M1+A1(cos(2\pi/24*t+phase1))+X(M*+A*(cos(2\pi/24*t+phase*)))+\epsilon}, where \eqn{\epsilon \sim N(0, \sigma^2)}; X = 0 for group 1 and X = 1 for group 2.
For group 2 the formula can be converted to \deqn{y = M2+A2(cos(2\pi/24*t+phase2))+\epsilon}
}
\examples{
#Simulate a time series data with 20 time points

B = 1
x1.time = runif(30, min = 0, max = 24)
x2.time = runif(30, min = 0, max = 24)
m1 = rnorm(B, 5); A1 = rnorm(B, 3); phase1 = runif(B, min = 0, max = 24); sigma = 1
m2 = m1; A2 = A1; phase2 = phase1

noise.mat1 = matrix(rnorm(B*30, 0, sigma), ncol = 30, nrow = B)
signal.mat1 = t(sapply(1:B, function(a){m1[a]+A1[a]*cos(2*pi/24*(x1.time+phase1[a]))}))
noise.mat2 = matrix(rnorm(B*30, 0, sigma), ncol = 30, nrow = B)
signal.mat2 = t(sapply(1:B, function(a){m2[a]+A2[a]*cos(2*pi/24*(x2.time+phase2[a]))}))

tod = c(x1.time, x2.time)
y = c(noise.mat1+signal.mat1, noise.mat2+signal.mat2)
group = c(rep(0, 30), rep(1, 30))

}
