% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CP_Rhythmicity.R
\name{CP_Rhythmicity}
\alias{CP_Rhythmicity}
\title{Title Rhythmicity Analysis for one data set}
\usage{
CP_Rhythmicity(
  x = list(data = data1, tod = tod, label = label),
  period = 24,
  method = "OLS",
  nPermutation = 1000,
  permutation.save = getwd(),
  permutation.file.label = "Group1",
  alpha = 0.05,
  p.adjust.method = "BH",
  parallel = TRUE,
  cores = 5
)
}
\arguments{
\item{x}{A list containing the following component: data: dataframe or matrix with rows of genes and columns of samples; tod: time of death or time of expression corresponding to the columns in data component; label: the gene names or other labels of the genes}

\item{period}{The length of the rhythmicity cycle. When it is 24, the signal is circadian.}

\item{method}{choose "OLS" or "permutation". OLS assumes Gaussian residual while permutation does not. Only OLS methods return confidence interval estimates. Notice that if permutation is selected, it is recommanded to use parallel computing.}

\item{nPermutation}{A numeric value. Number of permutation performer, only required when method = "permutation". The smallest possible p-value in the result will be 1/(n_gene*nPermutation), but larger nPermutation takes longer computing time.}

\item{permutation.save}{A local directory where you want to save the permutation result for future use. If "NULL" then no result will be saved.}

\item{permutation.file.label}{Special label for file name of permutation if you want to save the permutation file}

\item{alpha}{a number between 0 to 1. The critical level for the confidence interval}

\item{p.adjust.method}{choose valid input for p.adjust.method by checking p.adjust.methods for p.adjust() in R package \code{stat}}

\item{parallel}{TRUE/FALSE. If TRUE, parallel computing using mclapply will be used, which does not work on windows system.}

\item{cores}{number of cores used if using parallel computing}
}
\value{
A rhythm object, which is a list of original components in input x and rhythmicity results
}
\description{
This function fit the expression and time to this function \deqn{y = M+A(cos(2\pi/24*t+phase))+\epsilon}, where \eqn{\epsilon \sim N(0, \sigma^2)}.
}
\examples{
#Simulate a time series data with 20 time points and 100 genes

x.time = runif(20, min = 0, max = 24)
m = rnorm(100, 5); A = rnorm(100, 3); phase = runif(100, min = 0, max = 2*pi); sigma = 1
noise.mat = matrix(rnorm(100*20, 0, sigma), ncol = 20, nrow = 100)
signal.mat = t(sapply(1:100, function(a){m[a]+A[a]*cos(2*pi/24*x.time+phase[a])}))
x = list(data = noise.mat + signal.mat, tod = x.time, label = paste("gene", seq_len(100)))
x.Rhythm = CP_Rhythmicity(x, parallel = FALSE)


}
